+++ tomlkit/api.py
@@ -111,7 +111,7 @@ def table():  # type: () -> Table
 
 
 def inline_table():  # type: () -> InlineTable
-    return InlineTable(Container(), Trivia())
+    return InlineTable(Container(), Trivia(trail=""), new=True)
 
 
 def aot():  # type: () -> AoT